backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AmandaRibeiroLuz", email = "amanda.ribeiro.da.luz.201@gmail.com"},
]
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
